Changed the html file names in some places to better reflect what they are. Created an admin and moderator page which are functional. Admin can delete or penalize, moderator can penalize. Added in some user type checking for sensitive commands (like delete user).

+++ src/main/java/com/restResource/RestController.java
@@ -76,10 +76,9 @@ public class RestController {
 	@GET
 	@Produces(MediaType.APPLICATION_JSON)
 	@Path("/deleteUser/{usernameToDelete}")
-	public String deleteUser(@Context HttpServletRequest req, @PathParam("usernameToDelete") String usernameToDelete) {
-
-		if(dataController.verifyCredentials(em, req) == true) {	
-			
+	public String deleteUser(@Context HttpServletRequest req, @PathParam("usernameToDelete") String usernameToDelete) {		
+		
+		if(dataController.verifyCredentials(em, req) == true && dataController.getUserType(em, req).equals("{\"Result\":\"Admin\"}")) {	
 			try {
 				em.getTransaction().begin();
 				String statusMessage = dataController.deleteUser(em, usernameToDelete);
@@ -102,7 +101,7 @@ public class RestController {
 	@Path("/penalizeUser/{usernameToPenalize}")
 	public String penalizeUser(@Context HttpServletRequest req, @PathParam("usernameToPenalize") String usernameToPenalize) {
 
-		if(dataController.verifyCredentials(em, req) == true) {	
+		if(dataController.verifyCredentials(em, req) == true && (dataController.getUserType(em, req).equals("{\"Result\":\"Admin\"}") || dataController.getUserType(em, req).equals("{\"Result\":\"Moderator\"}"))) {	
 		
 			try {
 				em.getTransaction().begin();
